database "bsu_cool_chat_db" connected (#5)

+++ BSU_cool_chat/src/main/java/project/models/User.java
@@ -1,7 +1,7 @@
 package project.models;
 
 public class User {
-    private long id;
+    private int id;
     private String login;
 
     private String password;
@@ -14,7 +14,7 @@ public class User {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -34,7 +34,7 @@ public class User {
         this.password = password;
     }
 
-    public User(long id, String login, String password) {
+    public User(int id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
